Update README.md
My suggestion is to be explicit in the readme (I know it is given in the requirements file) about what's the lowest versioned Conda Python env that can correctly run the code. From trying out a few different Conda envs, it appears that Python 3.7.10 is just about as low as you can go and still run the code properly. At 3.6.10 Astropy throws an error. 

I would also like to suggest including the 'python' argument with 'conda create --name inspec-z'. Without the argument, Conda does not always create an env with an adequate Python version. For instance, on my system it creates in a Python 2.7.17 env, which of course is unable to run the code (without a bit of additional work). I don't know if I just happen to have a particularly quirky Python / Conda set up ...